requirements.txt
pymongo[srv]==4.6.1 

sentence-transformers==2.7.0 

nltk==3.9.1 

scikit-learn==1.4.0 

fastapi==0.110.0 

uvicorn==0.29.0 

groq==0.4.2 

python-dotenv==1.0.1 

pandas==2.2.1 

numpy==1.26.4 